# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Register a user and issue an initial token pair",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Authenticate with email and password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Rotate a refresh token and issue a new token pair",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Revoke a refresh token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/records": {
      "get": {
        "operationId": "RecordsController_list",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Category"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "updatedSince",
            "required": false,
            "in": "query",
            "description": "Only return records updated after this time.",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "schema": {
              "default": "false",
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 100,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordsPageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List the current user records",
        "tags": [
          "records"
        ]
      },
      "post": {
        "operationId": "RecordsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a record for the current user",
        "tags": [
          "records"
        ]
      }
    },
    "/api/records/{id}": {
      "get": {
        "operationId": "RecordsController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get one active record owned by the current user",
        "tags": [
          "records"
        ]
      },
      "patch": {
        "operationId": "RecordsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update an active record owned by the current user",
        "tags": [
          "records"
        ]
      },
      "delete": {
        "operationId": "RecordsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Soft-delete a record owned by the current user",
        "tags": [
          "records"
        ]
      }
    },
    "/api/health": {
      "get": {
        "operationId": "HealthController_getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Check whether the API process is running",
        "tags": [
          "health"
        ]
      }
    },
    "/api/health/db": {
      "get": {
        "operationId": "HealthController_getDatabaseHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseHealthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Check whether the database is reachable",
        "tags": [
          "health"
        ]
      }
    }
  },
  "info": {
    "title": "Neihao Records API",
    "description": "Authentication and synchronized dilemma-record APIs.",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72,
            "writeOnly": true,
            "example": "secure-pass-123"
          },
          "displayName": {
            "type": "string",
            "maxLength": 60,
            "example": "Lin"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "displayName": {
            "type": "object",
            "nullable": true,
            "example": "Lin"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "createdAt",
          "updatedAt"
        ]
      },
      "AuthTokens": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Short-lived JWT used as a Bearer access token."
          },
          "refreshToken": {
            "type": "string",
            "description": "Rotating JWT used only with the refresh endpoint."
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          },
          "tokens": {
            "$ref": "#/components/schemas/AuthTokens"
          }
        },
        "required": [
          "user",
          "tokens"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "oneOf": [
              {
                "type": "string",
                "example": "Invalid email or password."
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "email must be an email"
                ]
              }
            ]
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72,
            "writeOnly": true,
            "example": "secure-pass-123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token returned by register, login, or refresh."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Logged out."
          }
        },
        "required": [
          "message"
        ]
      },
      "Category": {
        "type": "string",
        "enum": [
          "WORK",
          "STUDY",
          "RELATIONSHIP",
          "CONSUMPTION",
          "CHOICE",
          "EMOTION",
          "OTHER"
        ]
      },
      "Emotion": {
        "type": "string",
        "enum": [
          "ANXIETY",
          "OVERTHINKING",
          "FEAR",
          "GRIEVANCE",
          "ANGER",
          "AVOIDANCE",
          "BLANK"
        ]
      },
      "TimeCost": {
        "type": "string",
        "enum": [
          "WITHIN_5_MINUTES",
          "WITHIN_30_MINUTES",
          "WITHIN_1_HOUR",
          "HALF_DAY",
          "OVER_1_DAY"
        ]
      },
      "WorthIt": {
        "type": "string",
        "enum": [
          "WORTH_IT",
          "NOT_WORTH_IT",
          "UNCLEAR"
        ]
      },
      "RecordSyncStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "DELETED"
        ]
      },
      "RecordResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "clientId": {
            "type": "object",
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "category": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Category"
              }
            ]
          },
          "emotions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Emotion"
            }
          },
          "emotionIntensity": {
            "type": "number",
            "minimum": 1,
            "maximum": 10
          },
          "decisionDifficulty": {
            "type": "number",
            "minimum": 1,
            "maximum": 10
          },
          "timeCost": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TimeCost"
              }
            ]
          },
          "thoughts": {
            "type": "string"
          },
          "finalDecision": {
            "type": "string"
          },
          "worthIt": {
            "allOf": [
              {
                "$ref": "#/components/schemas/WorthIt"
              }
            ]
          },
          "syncStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RecordSyncStatus"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "userId",
          "clientId",
          "title",
          "category",
          "emotions",
          "emotionIntensity",
          "decisionDifficulty",
          "timeCost",
          "thoughts",
          "finalDecision",
          "worthIt",
          "syncStatus",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "RecordsPageResponseDto": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecordResponseDto"
            }
          },
          "total": {
            "type": "number",
            "minimum": 0
          },
          "limit": {
            "type": "number",
            "minimum": 1,
            "maximum": 200
          },
          "offset": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "records",
          "total",
          "limit",
          "offset"
        ]
      },
      "CreateRecordDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Should I accept this task?",
            "description": "Stable client-generated ID."
          },
          "category": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Category"
              }
            ]
          },
          "emotions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Emotion"
            }
          },
          "emotionIntensity": {
            "type": "number",
            "minimum": 1,
            "maximum": 10,
            "example": 6
          },
          "decisionDifficulty": {
            "type": "number",
            "minimum": 1,
            "maximum": 10,
            "example": 7
          },
          "timeCost": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TimeCost"
              }
            ]
          },
          "thoughts": {
            "type": "string",
            "maxLength": 5000
          },
          "finalDecision": {
            "type": "string",
            "maxLength": 5000
          },
          "worthIt": {
            "allOf": [
              {
                "$ref": "#/components/schemas/WorthIt"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Original local creation time."
          }
        },
        "required": [
          "category",
          "emotions",
          "emotionIntensity",
          "decisionDifficulty",
          "timeCost",
          "worthIt"
        ]
      },
      "UpdateRecordDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "category": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Category"
              }
            ]
          },
          "emotions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Emotion"
            }
          },
          "emotionIntensity": {
            "type": "number",
            "minimum": 1,
            "maximum": 10
          },
          "decisionDifficulty": {
            "type": "number",
            "minimum": 1,
            "maximum": 10
          },
          "timeCost": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TimeCost"
              }
            ]
          },
          "thoughts": {
            "type": "string",
            "maxLength": 5000
          },
          "finalDecision": {
            "type": "string",
            "maxLength": 5000
          },
          "worthIt": {
            "allOf": [
              {
                "$ref": "#/components/schemas/WorthIt"
              }
            ]
          }
        }
      },
      "HealthResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          },
          "service": {
            "type": "string",
            "example": "neihao-records-backend"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "status",
          "service",
          "timestamp"
        ]
      },
      "DatabaseHealthResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          },
          "database": {
            "type": "string",
            "example": "reachable"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "status",
          "database",
          "timestamp"
        ]
      }
    }
  }
}